+++ flake.nix
@@ -15,7 +15,7 @@
         # Build from source — useful for development or patching.
         source = pkgs.buildGoModule rec {
           pname = "gorefact";
-          version = "0.0.12";
+          version = "0.0.13";
           src = ./.;
           go = pkgs.go;
           ldflags = [ "-s" "-w" "-X main.Version=${version}" ];